clean upload route

+++ backend/controllers/uploadController.js
@@ -1,21 +1,27 @@
-const multer = require('multer');
-const csv = require('csv-parser');
+const csvParser = require('csv-parser');
 const fs = require('fs');
 
-// Configuration de multer pour gérer les uploads de fichiers
-const upload = multer({ dest: 'uploads/' });
-
 const uploadCSV = (req, res) => {
+    if (!req.file) {
+        return res.status(400).send('Aucun fichier n\'a été téléversé');
+    }
+
     const results = [];
+
+    // Lire le fichier CSV et le parser
     fs.createReadStream(req.file.path)
-        .pipe(csv())
+        .pipe(csvParser())
         .on('data', (data) => results.push(data))
         .on('end', () => {
-            // Supprimez le fichier après le parsing
-            fs.unlinkSync(req.file.path);
-            // Traitez les données parsées ici
-            res.json(results);
+            try {
+                // Supprimez le fichier après le parsing
+                fs.unlinkSync(req.file.path);
+                // Traitez les données parsées ici
+                res.json(results);
+            } catch (error) {
+                res.status(500).json({ message: 'Erreur lors du traitement des données', error: error.message });
+            }
         });
 };
 
-module.exports = { upload, uploadCSV };
+module.exports = { uploadCSV };
